extra_addons/kids_club/requirements.txt
# Kids Club Module Dependencies
python-barcode>=0.13.1
Pillow>=8.0.0
